Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Geyner Rivera"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
tensorflow = "^2.11.0"
tensorflow-io-gcs-filesystem = "0.31.0"
uvicorn = "^0.29.0"
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
pillow = "^10.2.0"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
